looking for test

+++ internal/datamodels/models.go
@@ -35,7 +35,7 @@ type Withdraw struct {
 	Sum   float64 `json:"sum"`
 }
 type Accrual struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float32 `json:"accrual"`
 }
